Bash-Scripts/usefull/autoshutdown/requirements.txt
psutil
python-daemon